+++ library/common/jni/android_test_jni_interface.cc
@@ -1,4 +1,4 @@
-#include "jni_import.h"
+#include "library/common/jni/import/jni_import.h"
 #include "library/common/jni/jni_support.h"
 #include "library/common/jni/jni_utility.h"
 #include "library/common/main_interface.h"
