+++ library/common/jni/jni_support.h
@@ -1,4 +1,4 @@
-#include "jni_import.h"
+#include "library/common/jni/import/jni_import.h"
 
 // NOLINT(namespace-envoy)
