+++ library/common/jni/jni_version.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "jni_import.h"
+#include "library/common/jni/import/jni_import.h"
 
 // NOLINT(namespace-envoy)
